Implement DOMpurify to all the files that need it and break the function to separate file.
